requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
scipy>=1.11.0
numpy>=1.24.0
plotly>=5.17.0
pytz>=2023.3
